requirements.txt
torch>=2.0.0
transformers>=4.45.0
peft>=0.12.0
bitsandbytes>=0.43.0
accelerate>=0.34.0
sentencepiece
pillow
tensorboard